refactor: merge vision proxy processor into vision proxy module.
