+++ internal/server/module/visionproxy/vision_proxy.go
@@ -1,4 +1,4 @@
-package processor
+package visionproxy
 
 import (
 	"context"
