+++ internal/server/module/visionproxy/vision_proxy_tool_result_test.go
@@ -1,4 +1,4 @@
-package processor
+package visionproxy
 
 import (
 	"strings"
